# cart-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Cart API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /customer/{customerId}/cart/product/{productId}:
    x-swagger-router-controller: cart
    put:
      operationId: update
      description: add a new product to the cart
      parameters:
        - name: customerId
          in: path
          description: ID of customer which cart this is
          required: true
          type: string
        - name: productId
          in: path
          description: ID of product in cart
          required: true
          type: string
        - name: product
          description: product properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/editcartproduct"
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delProduct
      description: remove a product from cart
      parameters:
        - name: customerId
          in: path
          description: ID of customer which cart this is
          required: true
          type: string
        - name: productId
          in: path
          description: ID of product in cart
          required: true
          type: string
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  editcartproduct:
    type: object
    properties: 
      amount:
        type: number
      price: 
        type: number
...
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
...
  ErrorResponse:
    required:
      - code
      - message
    properties:
      code: 
        type: string
      message:
        type: string